requirements.txt
fastapi
uvicorn
pandas
numpy
pydantic
pymysql
huggingface_hub
python-dotenv
langchain
langchain_huggingface
cryptography